Scope logging to differential.* and extract helpers from main.py
Set root logger to WARNING and only promote the differential logger,
so new dependencies are automatically silenced without per-library
overrides. Move source-file helpers to differential.sources and
create_root_question to differential.orchestrator — fixes a broken
import (from llm import run_llm) and makes generate_source_summary
properly async.
